LPD-37449 Export normalizeSourceItem to be reused
