+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/getNormalizedDragItems.ts
@@ -5,9 +5,7 @@
 
 import {FragmentLayoutDataItem} from '../../types/layout_data/FragmentLayoutDataItem';
 import {FragmentEntryLinkMap} from '../actions/addFragmentEntryLinks';
-import {LAYOUT_DATA_ITEM_TYPES} from '../config/constants/layoutDataItemTypes';
-import selectLayoutDataItemLabel from '../selectors/selectLayoutDataItemLabel';
-import isItemWidget from './isItemWidget';
+import normalizeSourceItem from './normalizeSourceItem';
 
 import type {LayoutData} from '../../types/layout_data/LayoutData';
 
@@ -17,36 +15,6 @@ type SourceItem = FragmentLayoutDataItem & {
 	isWidget: boolean;
 	name: string;
 };
-
-function normalizeSourceItem(
-	item: FragmentLayoutDataItem,
-	layoutData: LayoutData,
-	fragmentEntryLinks: FragmentEntryLinkMap
-) {
-	const isWidget = isItemWidget(item, fragmentEntryLinks);
-	const name = selectLayoutDataItemLabel(
-		{
-			fragmentEntryLinks,
-			layoutData,
-		},
-		item
-	);
-
-	let fragmentEntryLink = null;
-
-	if (item.type === LAYOUT_DATA_ITEM_TYPES.fragment) {
-		fragmentEntryLink =
-			fragmentEntryLinks[item.config?.fragmentEntryLinkId];
-	}
-
-	return {
-		...item,
-		fieldTypes: fragmentEntryLink?.fieldTypes ?? [],
-		fragmentEntryType: fragmentEntryLink?.fragmentEntryType ?? null,
-		isWidget,
-		name,
-	};
-}
 
 export default function getNormalizedDragItems(
 	sourceItem: SourceItem,
